+++ bld/commonui/h/pushwin.h
@@ -33,13 +33,13 @@
 #define _PUSHWIN_H_INCLUDED
 
 typedef struct {
-    WORD                id;
+    int                 id;
     HFONT               font;
     HWND                parent;
     char                str[1];         /* dynamic array */
 } PushWinInfo;
 
 bool    RegPushWin( HANDLE instance );
-HWND    CreatePushWin( HWND parent, char *txt, WORD id, HFONT font, HANDLE inst );
+HWND    CreatePushWin( HWND parent, char *txt, int id, HFONT font, HANDLE inst );
 
 #endif /* _PUSHWIN_H_INCLUDED */
